chore: replace the contents of ShaderityUtility with NewShaderityUtility
